Fix security vulnerability and cleanup deps (#11)
* Fix security venerability

During install, npm warned against the following venerability
https://github.com/advisories/GHSA-67mh-4wv8-2f99 . This commit bumps
the effected esbuild version and vitest, which depends on it.

* Bump and Remove Unused Dependencies, add Formatting

This commit updates prettier, so it now respects and therefore ignores
files in the folder dist.  Furthermore, arg, npm-run-all and globby are
not used and only pull in unused code and therefore cause a wider attack
surface on developer machines

* Format Code, and Fix Spacing and Name in bench.js

The spacing in scripts/bench.js was off, leading to an unconsitent
visual output when running the benchmark.  Furthermore, the benchmark
was still named ultraflags.  Prettier was used to format the rest of the
code

* Revert "Format Code, and Fix Spacing and Name in bench.js"

This reverts commit 4f460e558ed645f91e0c9b3140ee836d3a1ca93f.

* Fix Spacing and Library Name in bench.js

The spacing in scripts/bench.js was off, leading to an unconsitent
visual output when running the benchmark.  Furthermore, the benchmark
was still named ultraflags.